requirements.txt
aiohttp>=3.9.0
beautifulsoup4>=4.12.0
python-telegram-bot>=21.0
openai>=1.0.0
tavily-python>=0.3.0
duckduckgo_search>=6.0.0
python-dotenv>=1.0.0
